Revert "[Refacto pour fonctionnel]"
