+++ lib/state/retrieve_content_by_user_id/retrieve_content_by_user_id_cubit.dart
@@ -21,38 +21,4 @@ class RetrieveContentByUserIdCubit extends Cubit<RetrieveContentByUserIdState> {
       emit(RetrieveContentByUserIdState.loaded(r));
     });
   }
-
-  void like(String publicationId) {
-    state.maybeWhen(
-      orElse: () {},
-      loaded: (contents) {
-        final loadedState = state as _Loaded;
-        emit(loadedState.copyWith(
-          publications: loadedState.publications.map((content) {
-            if (content.id == publicationId) {
-              return content.copyWith(isLike: true, isDislike: false);
-            }
-            return content;
-          }).toList(),
-        ));
-      },
-    );
-  }
-
-  void dislike(String publicationId) {
-    state.maybeWhen(
-      orElse: () {},
-      loaded: (contents) {
-        final loadedState = state as _Loaded;
-        emit(loadedState.copyWith(
-          publications: loadedState.publications.map((content) {
-            if (content.id == publicationId) {
-              return content.copyWith(isLike: false, isDislike: true);
-            }
-            return content;
-          }).toList(),
-        ));
-      },
-    );
-  }
 }
