+++ lib/ui/pages/account/widgets/account_loaded.dart
@@ -4,8 +4,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:getwidget/components/avatar/gf_avatar.dart';
 import 'package:getwidget/components/button/gf_button.dart';
 import 'package:getwidget/shape/gf_button_shape.dart';
-import '../../../../ioc_container.dart';
-import '../../../../state/like/like_cubit.dart';
 import '../../../../state/retrieve_follow_by_follower_id/retrieve_follow_by_follower_id_cubit.dart';
 
 import '../../../../domain/entities/user.dart';
@@ -14,7 +12,6 @@ import '../../../../state/retrieve_follow_by_user_id/retrieve_follow_by_user_id_
 import '../../../default_pictures.dart';
 import '../../../widgets/loading.dart';
 import '../../home/widgets/home_loaded.dart';
-import '../../home/widgets/publications_loaded.dart';
 
 class AccountLoaded extends StatelessWidget {
   final User user;
@@ -157,26 +154,7 @@ class AccountLoaded extends StatelessWidget {
                 return const Loading();
               },
               orElse: () => const Loading(),
-              loaded: (publications) => Container(
-                alignment: Alignment.center,
-                width: double.infinity,
-                height: MediaQuery.of(context).size.height,
-                child: Column(
-                  children: [
-                    Flexible(
-                      child: BlocProvider(
-                          create: (context) => LikeCubit(
-                            sl(),
-                            null,
-                            context.read<RetrieveContentByUserIdCubit>(),
-                            sl(),
-                          ),
-                          child: PublicationsLoaded(publications: publications)
-                      ),
-                    ),
-                  ],
-                ),
-              ),
+              loaded: (publications) => HomeLoaded(publications: publications),
             );
           },
         ),
